pyproject.toml
[tool.poetry]
name = "ze-delivery-code-challenge"
version = "0.1.0"
description = ""
authors = ["Joao Pedro Watson <joaopedrowatson@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.0.3"
tinydb = "^4.7.0"
python-dotenv = "^0.19.2"
apifairy = "^0.9.1"
alchemical = "^0.5.1"
Flask-Cors = "^3.0.10"
Flask-Migrate = "^3.1.0"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.28.0"

[tool.poetry.dev-dependencies]
blue = "^0.8.0"
isort = "^5.10.1"
prospector = "^1.7.7"
uvicorn = {extras = ["starndard"], version = "^0.17.6"}
httpie = "^3.1.0"
mkdocs = "^1.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
